// packages/core/protocol/proto/decide/v1alpha1/decide_pb.d.ts
export declare enum EmailType {
  /**
   * The email type is unspecified. This should not be used, but is here to
   * conform to the gRPC best practices.
   *
   * @generated from enum value: EMAIL_TYPE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * The email address is disposable, which means it's registered to a service
   * that allows throwaway email addresses. Although these are sometimes used
   * for privacy, they are also often used for spam signups or fraudulent
   * activity when combined with a transaction e.g. attempting to use a credit
   * card. We recommend blocking these in higher risk scenarios.
   *
   * @generated from enum value: EMAIL_TYPE_DISPOSABLE = 1;
   */
  DISPOSABLE = 1,

  /**
   * The email address is registered to a free email service. These are very
   * common, such as GMail or Yahoo Mail, so we do not recommend blocking
   * these. However, you may wish to flag these for review the first time they
   * attempt a transaction.
   *
   * @generated from enum value: EMAIL_TYPE_FREE = 2;
   */
  FREE = 2,

  /**
   * This email address is registered to a domain name which has no MX records
   * configured. This means it cannot receive email. We recommend blocking
   * these.
   *
   * @generated from enum value: EMAIL_TYPE_NO_MX_RECORDS = 3;
   */
  NO_MX_RECORDS = 3,

  /**
   * This email has no Gravatar attached to the email from
   * https://gravatar.com which makes it slightly less likely to be a valid
   * signup. We recommend using this as part of your own risk scoring or
   * manually reviewing these signups.
   *
   * @generated from enum value: EMAIL_TYPE_NO_GRAVATAR = 4;
   */
  NO_GRAVATAR = 4,

  /**
   * This email was specified in an invalid format.
   *
   * @generated from enum value: EMAIL_TYPE_INVALID = 5;
   */
  INVALID = 5,
}

/**
 * The mode to run in. This can be either `DRY_RUN` or `LIVE`. In `DRY_RUN`
 * mode, all requests will be allowed and you can review what the action would
 * have been from your dashboard. In `LIVE` mode, requests will be allowed,
 * challenged or blocked based on the returned decision.
 *
 * @generated from enum proto.decide.v1alpha1.Mode
 */
export declare enum Mode {
  /**
   * The mode is unspecified. This should not be used, but is here to conform
   * to the gRPC best practices.
   *
   * @generated from enum value: MODE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * In `DRY_RUN` mode, all requests will be allowed and you can review what
   * the action would have been from your dashboard.
   *
   * @generated from enum value: MODE_DRY_RUN = 1;
   */
  DRY_RUN = 1,

  /**
   * In `LIVE` mode, requests will be allowed, challenged or blocked based on
   * the returned decision.
   *
   * @generated from enum value: MODE_LIVE = 2;
   */
  LIVE = 2,
}
// ...
export declare enum Conclusion {
  /**
   * The conclusion is unspecified. This should not be used, but is here to
   * conform to the gRPC best practices.
   *
   * @generated from enum value: CONCLUSION_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * The request should be allowed.
   *
   * @generated from enum value: CONCLUSION_ALLOW = 1;
   */
  ALLOW = 1,

  /**
   * The request should be blocked.
   *
   * @generated from enum value: CONCLUSION_DENY = 2;
   */
  DENY = 2,

  /**
   * The request should be challenged.
   *
   * @generated from enum value: CONCLUSION_CHALLENGE = 3;
   */
  CHALLENGE = 3,

  /**
   * The request errored.
   *
   * @generated from enum value: CONCLUSION_ERROR = 4;
   */
  ERROR = 4,
}
// ...
export declare enum RateLimitAlgorithm {
  /**
   * @generated from enum value: RATE_LIMIT_ALGORITHM_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: RATE_LIMIT_ALGORITHM_TOKEN_BUCKET = 1;
   */
  TOKEN_BUCKET = 1,

  /**
   * @generated from enum value: RATE_LIMIT_ALGORITHM_FIXED_WINDOW = 2;
   */
  FIXED_WINDOW = 2,

  /**
   * @generated from enum value: RATE_LIMIT_ALGORITHM_SLIDING_WINDOW = 3;
   */
  SLIDING_WINDOW = 3,
}
// ...
export declare class Reason extends Message<Reason> {
  /**
   * @generated from oneof proto.decide.v1alpha1.Reason.reason
   */
  reason: {
    /**
     * Contains details about the rate limit when the decision was made
     * based on a rate limit rule.
     *
     * @generated from field: proto.decide.v1alpha1.RateLimitReason rate_limit = 1;
     */
    value: RateLimitReason;
    case: "rateLimit";
  } | {
    /**
     * Contains details about the edge rules which were triggered when the
     * decision was made based on an edge rule.
     *
     * @generated from field: proto.decide.v1alpha1.EdgeRuleReason edge_rule = 2;
     */
    value: EdgeRuleReason;
    case: "edgeRule";
  } | {
    /**
     * Contains details about why the request was considered a bot when the
     * decision was made based on a bot rule.
     *
     * @generated from field: proto.decide.v1alpha1.BotReason bot = 3;
     */
    value: BotReason;
    case: "bot";
  } | {
    /**
     * Contains details about why Snipkit Shield was triggered.
     *
     * @generated from field: proto.decide.v1alpha1.ShieldReason shield = 4;
     */
    value: ShieldReason;
    case: "shield";
  } | {
    /**
     * Contains details about the email when the decision was made based on
     * an email rule.
     *
     * @generated from field: proto.decide.v1alpha1.EmailReason email = 5;
     */
    value: EmailReason;
    case: "email";
  } | {
    /**
     * Contains details about the error decision when an error occurred.
     *
     * @generated from field: proto.decide.v1alpha1.ErrorReason error = 6;
     */
    value: ErrorReason;
    case: "error";
  } | {
    /**
     * Contains details about sensitive info identified in the body of the
     * request if the sensitive info rule is configured.
     *
     * @generated from field: proto.decide.v1alpha1.SensitiveInfoReason sensitive_info = 7;
     */
    value: SensitiveInfoReason;
    case: "sensitiveInfo";
  } | {
    /**
     * Contains details about why the request was considered a bot when the
     * decision was made based on a bot (v2) rule.
     *
     * @generated from field: proto.decide.v1alpha1.BotV2Reason bot_v2 = 8;
     */
    value: BotV2Reason;
    case: "botV2";
  } | { case: undefined; value?: undefined };

  constructor(data?: PartialMessage<Reason>);

  static readonly runtime: typeof proto3;
  static readonly typeName = "proto.decide.v1alpha1.Reason";
  static readonly fields: FieldList;

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Reason;

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Reason;

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Reason;

  static equals(a: Reason | PlainMessage<Reason> | undefined, b: Reason | PlainMessage<Reason> | undefined): boolean;
}
// ...
export declare class Rule extends Message<Rule> {
  /**
   * @generated from oneof proto.decide.v1alpha1.Rule.rule
   */
  rule: {
    /**
     * @generated from field: proto.decide.v1alpha1.RateLimitRule rate_limit = 1;
     */
    value: RateLimitRule;
    case: "rateLimit";
  } | {
    /**
     * @generated from field: proto.decide.v1alpha1.BotRule bots = 2;
     */
    value: BotRule;
    case: "bots";
  } | {
    /**
     * @generated from field: proto.decide.v1alpha1.EmailRule email = 3;
     */
    value: EmailRule;
    case: "email";
  } | {
    /**
     * @generated from field: proto.decide.v1alpha1.ShieldRule shield = 4;
     */
    value: ShieldRule;
    case: "shield";
  } | {
    /**
     * @generated from field: proto.decide.v1alpha1.SensitiveInfoRule sensitive_info = 5;
     */
    value: SensitiveInfoRule;
    case: "sensitiveInfo";
  } | {
    /**
     * @generated from field: proto.decide.v1alpha1.BotV2Rule bot_v2 = 6;
     */
    value: BotV2Rule;
    case: "botV2";
  } | { case: undefined; value?: undefined };

  constructor(data?: PartialMessage<Rule>);

  static readonly runtime: typeof proto3;
  static readonly typeName = "proto.decide.v1alpha1.Rule";
  static readonly fields: FieldList;

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Rule;

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Rule;

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Rule;

  static equals(a: Rule | PlainMessage<Rule> | undefined, b: Rule | PlainMessage<Rule> | undefined): boolean;
}

/**
 * @generated from message proto.decide.v1alpha1.RuleResult
 */
export declare class RuleResult extends Message<RuleResult> {
  /**
   * The stable, deterministic, and unique identifier of the rule that
   * generated this result.
   *
   * @generated from field: string rule_id = 1;
   */
  ruleId: string;

  /**
   * The rule run state
   *
   * @generated from field: proto.decide.v1alpha1.RuleState state = 2;
   */
  state: RuleState;

  /**
   * The conclusion determined by the rule.
   *
   * @generated from field: proto.decide.v1alpha1.Conclusion conclusion = 3;
   */
  conclusion: Conclusion;

  /**
   * The reason for the conclusion.
   *
   * @generated from field: proto.decide.v1alpha1.Reason reason = 4;
   */
  reason?: Reason;

  /**
   * The duration in seconds this result should be considered valid, also
   * known as time-to-live.
   *
   * @generated from field: uint32 ttl = 5;
   */
  ttl: number;

  /**
   * The fingerprint calculated for this rule, which can be used to cache the
   * result for the amount of time specified by `ttl`.
   *
   * @generated from field: string fingerprint = 6;
   */
  fingerprint: string;

  constructor(data?: PartialMessage<RuleResult>);

  static readonly runtime: typeof proto3;
  static readonly typeName = "proto.decide.v1alpha1.RuleResult";
  static readonly fields: FieldList;

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): RuleResult;

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): RuleResult;

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): RuleResult;

  static equals(a: RuleResult | PlainMessage<RuleResult> | undefined, b: RuleResult | PlainMessage<RuleResult> | undefined): boolean;
}
// ...
export declare class Decision extends Message<Decision> {
  /**
   * The decision ID. This is a unique identifier for the decision which can
   * be used to search for the request details in the Snipkit dashboard.
   *
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * Snipkit's conclusion for the request based on our analysis.
   *
   * @generated from field: proto.decide.v1alpha1.Conclusion conclusion = 2;
   */
  conclusion: Conclusion;

  /**
   * The reason for the decision.
   *
   * @generated from field: proto.decide.v1alpha1.Reason reason = 3;
   */
  reason?: Reason;

  /**
   * The outcome of each rule taken into consideration for the decision.
   *
   * @generated from field: repeated proto.decide.v1alpha1.RuleResult rule_results = 4;
   */
  ruleResults: RuleResult[];

  /**
   * The duration in seconds this decision should be considered valid, also
   * known as time-to-live.
   *
   * @generated from field: uint32 ttl = 5;
   */
  ttl: number;

  /**
   * Details about the IP address that informed our conclusion.
   *
   * @generated from field: proto.decide.v1alpha1.IpDetails ip_details = 6;
   */
  ipDetails?: IpDetails;

  constructor(data?: PartialMessage<Decision>);

  static readonly runtime: typeof proto3;
  static readonly typeName = "proto.decide.v1alpha1.Decision";
  static readonly fields: FieldList;

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Decision;

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Decision;

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Decision;

  static equals(a: Decision | PlainMessage<Decision> | undefined, b: Decision | PlainMessage<Decision> | undefined): boolean;
}